Remove duplicate event model
