+++ app/src/main/java/org/hackillinois/android/database/Database.kt
@@ -5,7 +5,7 @@ import android.arch.persistence.room.RoomDatabase
 import org.hackillinois.android.database.dao.*
 import org.hackillinois.android.database.entity.*
 
-@Database(entities = [QR::class, Attendee::class, User::class, Event::class, Roles::class], version = 5)
+@Database(entities = [QR::class, Attendee::class, User::class, Event::class, Roles::class], version = 6)
 abstract class Database : RoomDatabase() {
     abstract fun qrDao(): QRDao
     abstract fun attendeeDao(): AttendeeDao
