+++ app/src/main/java/org/hackillinois/android/database/entity/Event.kt
@@ -14,7 +14,9 @@ data class Event(
     val description: String,
     val startTime: Long,
     val endTime: Long,
-    val locations: List<EventLocation>
+    val locations: List<EventLocation>,
+    val sponsor: String,
+    val eventType: String
 ) : BaseEntity() {
     
     fun getStartTimeMs() = startTime * 1000L
@@ -58,3 +60,6 @@ data class EventLocation (
     val latitude: Double,
     val longitude: Double
 )
+
+val SiebelCenter = EventLocation("Siebel Center", 40.1138, -88.2249)
+val EceBuilding = EventLocation("ECE Building", 40.1148, -88.2280)
